requirements.txt
# Core deep learning
torch>=2.0.0
torchaudio>=2.0.0
pytorch-lightning>=2.0.0

# Audio processing
librosa>=0.10.0
soundfile>=0.12.1

# Data processing and analysis
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Configuration and logging
pyyaml>=6.0
tensorboard>=2.13.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
reportlab>=4.0.0

# Development tools
pytest>=7.3.1
black>=23.3.0
flake8>=6.0.0
mypy>=1.3.0

# Optional: HiFi-GAN vocoder (if needed)
# Install from source or use provided checkpoint
# git+https://github.com/jik876/hifi-gan.git 